Add new functionalities. Re-write some of the code.

+++ funcții.php
@@ -10,29 +10,64 @@ function selectează($tabel = 'elevi' , $coloană = '', $valoare = ''){
     $comandă -> bindParam (1, $valoare);
   }
   $comandă -> execute();
-  return $comandă -> fetchAll(PDO::FETCH_NUM);
+  return $comandă -> fetchAll(PDO::FETCH_ASSOC);
 }
 
-function printează_tabela($date_sql){
-  $lungime_date_sql = count($date_sql);
-  for ($i = 0; $i < $lungime_date_sql; $i++){
+function selectează_note($tabel = 'elevi' , $coloană = '', $valoare = ''){
+  global $con;
+  $comandă = $con->prepare("SELECT note.id,materie,notă FROM note INNER JOIN
+  materii ON materii.id = note.id_materie where id_elev = ?;");
+  $comandă -> bindParam (1, $valoare);
+  $comandă -> execute();
+  return $comandă -> fetchAll(PDO::FETCH_ASSOC);
+}
+
+function printează($date_sql, $numar_randuri, $harta_ieșire=[]){
+  for ($rand = 0; $rand < $numar_randuri; $rand++){
     echo "<tr>";
-    for ($j = 0; $j < count($date_sql[$i]); $j++){
+    foreach ($date_sql[$rand] as $cheie => $valoare){
+      if ( array_key_exists($cheie, $harta_ieșire)){
+        continue;
+      }
       echo "<td>";
-      echo $date_sql[$i][$j];
+      echo $valoare;
       echo "</td>";
     }
     echo "</tr>";
   }
 }
 
-function inserează($tabel, $var_1, $var_2, $var_3){
-  $vector = [$var_1, $var_2, intval($var_3)];
+function printează_note($date_sql, $numar_randuri, $harta_ieșire=[]){
+  for ($rand = 0; $rand < $numar_randuri; $rand++){
+    echo "<tr>";
+    foreach ($date_sql[$rand] as $cheie => $valoare){
+      if ( $cheie != 'prenume') 
+        echo "<td>";
+      echo $valoare." ";
+      if ( $cheie != 'nume')
+        echo "</td>";
+    }
+    echo "</tr>";
+  }
+}
+
+function inserează($tabel, $var_1, $var_2 = '', $var_3 = ''){
+  $vector = [$var_1];
   foreach ($vector as &$membru) {
     $membru = preg_replace("/[^a-zA-Z 0-9ăîâșțĂÎÂȘȚ]+/", "", $membru );
   }
   global $con;
-  $comandă = $con->prepare("INSERT INTO $tabel VALUES (DEFAULT, ?, ?, ?)");
+  $text_comandă = "INSERT INTO $tabel VALUES (DEFAULT, ?";
+  if ($var_2) {
+    array_push($vector,$var_2);
+    $text_comandă = $text_comandă . ", ?";
+    if ($var_3) {
+      array_push($vector, $var_3);
+      $text_comandă = $text_comandă . ", ?";
+    }
+  }
+  $text_comandă = $text_comandă . ")";
+  $comandă = $con->prepare($text_comandă);
   $comandă -> execute($vector);
 }
 
@@ -40,5 +75,13 @@ function șterge($tabel, $var1) {
   global $con;
   $comandă = $con->prepare("DELETE FROM $tabel WHERE id = ?");
   $comandă -> execute([$var1]);
+}
+
+function print_comandă_creare($tabelă){
+  global $con;
+  $comandă = $con->prepare("SHOW CREATE TABLE $tabelă");
+  $comandă -> execute();
+  return $comandă -> fetchAll()[0]['Create Table'];
+  return $comandă;
 }
 ?>
